Create Grades djangoapp
https://openedx.atlassian.net/browse/TNL-5008
