+++ lms/djangoapps/certificates/tests/test_cert_management.py
@@ -1,6 +1,5 @@
 """Tests for the resubmit_error_certificates management command. """
 import ddt
-from contextlib import contextmanager
 from django.core.management.base import CommandError
 from nose.plugins.attrib import attr
 from django.test.utils import override_settings
@@ -12,6 +11,7 @@ from opaque_keys.edx.locator import CourseLocator
 from badges.events.course_complete import get_completion_badge
 from badges.models import BadgeAssertion
 from badges.tests.factories import BadgeAssertionFactory, CourseCompleteImageConfigurationFactory
+from lms.djangoapps.grades.tests.utils import mock_passing_grade
 from xmodule.modulestore.tests.django_utils import ModuleStoreTestCase
 from xmodule.modulestore.tests.factories import CourseFactory, check_mongo_calls, ItemFactory
 from student.tests.factories import UserFactory, CourseEnrollmentFactory
@@ -247,7 +247,7 @@ class UngenerateCertificatesTest(CertificateManagementTest):
         mock_send_to_queue.return_value = (0, "Successfully queued")
         key = self.course.location.course_key
         self._create_cert(key, self.user, CertificateStatuses.unavailable)
-        with self._mock_passing_grade():
+        with mock_passing_grade():
             self._run_command(
                 course=unicode(key), noop=False, insecure=True, force=False
             )
@@ -257,11 +257,3 @@ class UngenerateCertificatesTest(CertificateManagementTest):
             course_id=key
         )
         self.assertEqual(certificate.status, CertificateStatuses.generating)
-
-    @contextmanager
-    def _mock_passing_grade(self):
-        """Mock the grading function to always return a passing grade. """
-        symbol = 'courseware.grades.grade'
-        with patch(symbol) as mock_grade:
-            mock_grade.return_value = {'grade': 'Pass', 'percent': 0.75}
-            yield
