+++ lms/djangoapps/certificates/tests/test_webview_views.py
@@ -21,6 +21,7 @@ from lms.djangoapps.badges.tests.factories import (
     CourseCompleteImageConfigurationFactory,
     BadgeClassFactory,
 )
+from lms.djangoapps.grades.tests.utils import mock_passing_grade
 from openedx.core.lib.tests.assertions.events import assert_event_matches
 from student.tests.factories import UserFactory, CourseEnrollmentFactory
 from student.roles import CourseStaffRole
@@ -836,8 +837,7 @@ class CertificatesViewsTests(CommonCertificatesTestCase):
         request_certificate_url = reverse('certificates.views.request_certificate')
         with patch('capa.xqueue_interface.XQueueInterface.send_to_queue') as mock_queue:
             mock_queue.return_value = (0, "Successfully queued")
-            with patch('courseware.grades.grade') as mock_grade:
-                mock_grade.return_value = {'grade': 'Pass', 'percent': 0.75}
+            with mock_passing_grade():
                 response = self.client.post(request_certificate_url, {'course_id': unicode(self.course.id)})
                 self.assertEqual(response.status_code, 200)
                 response_json = json.loads(response.content)
