+++ lms/djangoapps/gating/api.py
@@ -9,6 +9,7 @@ from django.contrib.auth.models import User
 from xmodule.modulestore.django import modulestore
 from milestones import api as milestones_api
 from openedx.core.lib.gating import api as gating_api
+from lms.djangoapps.grades.module_grades import get_module_score
 
 
 log = logging.getLogger(__name__)
@@ -64,7 +65,6 @@ def evaluate_prerequisite(course, prereq_content_key, user_id):
 
             gated_content = gated_content_milestones.get(prereq_milestone['id'])
             if gated_content:
-                from courseware.grades import get_module_score
                 user = User.objects.get(id=user_id)
                 score = get_module_score(user, course, sequential) * 100
                 for milestone in gated_content:
