+++ lms/djangoapps/gating/signals.py
@@ -4,7 +4,7 @@ Signal handlers for the gating djangoapp
 from django.dispatch import receiver
 from opaque_keys.edx.keys import CourseKey, UsageKey
 from xmodule.modulestore.django import modulestore
-from courseware.models import SCORE_CHANGED
+from lms.djangoapps.grades.signals import SCORE_CHANGED
 from gating import api as gating_api
