+++ lms/djangoapps/gating/tests/test_api.py
@@ -136,7 +136,7 @@ class TestEvaluatePrerequisite(GatingTestCase, MilestonesTestCaseMixin):
         gating_api.set_required_content(self.course.id, self.seq2.location, self.seq1.location, min_score)
         self.prereq_milestone = gating_api.get_gating_milestone(self.course.id, self.seq1.location, 'fulfills')
 
-    @patch('courseware.grades.get_module_score')
+    @patch('gating.api.get_module_score')
     @data((.5, True), (1, True), (0, False))
     @unpack
     def test_min_score_achieved(self, module_score, result, mock_module_score):
@@ -149,7 +149,7 @@ class TestEvaluatePrerequisite(GatingTestCase, MilestonesTestCaseMixin):
         self.assertEqual(milestones_api.user_has_milestone(self.user_dict, self.prereq_milestone), result)
 
     @patch('gating.api.log.warning')
-    @patch('courseware.grades.get_module_score')
+    @patch('gating.api.get_module_score')
     @data((.5, False), (1, True))
     @unpack
     def test_invalid_min_score(self, module_score, result, mock_module_score, mock_log):
@@ -162,21 +162,21 @@ class TestEvaluatePrerequisite(GatingTestCase, MilestonesTestCaseMixin):
         self.assertEqual(milestones_api.user_has_milestone(self.user_dict, self.prereq_milestone), result)
         self.assertTrue(mock_log.called)
 
-    @patch('courseware.grades.get_module_score')
+    @patch('gating.api.get_module_score')
     def test_orphaned_xblock(self, mock_module_score):
         """ Test test_orphaned_xblock """
 
         evaluate_prerequisite(self.course, self.prob2.location, self.user.id)
         self.assertFalse(mock_module_score.called)
 
-    @patch('courseware.grades.get_module_score')
+    @patch('gating.api.get_module_score')
     def test_no_prerequisites(self, mock_module_score):
         """ Test test_no_prerequisites """
 
         evaluate_prerequisite(self.course, self.prob1.location, self.user.id)
         self.assertFalse(mock_module_score.called)
 
-    @patch('courseware.grades.get_module_score')
+    @patch('gating.api.get_module_score')
     def test_no_gated_content(self, mock_module_score):
         """ Test test_no_gated_content """
