+++ lms/djangoapps/instructor/offline_gradecalc.py
@@ -10,9 +10,10 @@ import json
 import time
 
 from json import JSONEncoder
-from courseware import grades, models
+from courseware import models
 from courseware.courses import get_course_by_id
 from django.contrib.auth.models import User
+from lms.djangoapps.grades import course_grades
 from opaque_keys import OpaqueKey
 from opaque_keys.edx.keys import UsageKey
 from xmodule.graders import Score
@@ -50,7 +51,7 @@ def offline_grade_calculation(course_key):
         request.user = student
         request.session = {}
 
-        gradeset = grades.grade(student, course, keep_raw_scores=True)
+        gradeset = course_grades.summary(student, course, keep_raw_scores=True)
         # Convert Score namedtuples to dicts:
         totaled_scores = gradeset['totaled_scores']
         for section in totaled_scores:
@@ -89,7 +90,7 @@ def student_grades(student, request, course, keep_raw_scores=False, use_offline=
     as use_offline.  If use_offline is True then this will look for an offline computed gradeset in the DB.
     '''
     if not use_offline:
-        return grades.grade(student, course, keep_raw_scores=keep_raw_scores)
+        return course_grades.summary(student, course, keep_raw_scores=keep_raw_scores)
 
     try:
         ocg = models.OfflineComputedGrade.objects.get(user=student, course_id=course.id)
