+++ lms/djangoapps/mobile_api/users/tests.py
@@ -25,6 +25,7 @@ from courseware.access_response import (
     VisibilityError,
 )
 from course_modes.models import CourseMode
+from lms.djangoapps.grades.tests.utils import mock_passing_grade
 from openedx.core.lib.courses import course_image_url
 from student.models import CourseEnrollment
 from util.milestones_helpers import set_prerequisite_courses
@@ -247,8 +248,7 @@ class TestUserEnrollmentApi(UrlResetMixin, MobileAPITestCase, MobileAuthUserTest
         self.course.cert_html_view_enabled = True
         self.store.update_item(self.course, self.user.id)
 
-        with patch('courseware.grades.grade') as mock_grade:
-            mock_grade.return_value = {'grade': 'Pass', 'percent': 0.75}
+        with mock_passing_grade():
             generate_user_certificates(self.user, self.course.id)
 
         response = self.api_response()
