+++ openedx/core/djangoapps/signals/signals.py
@@ -5,7 +5,7 @@ This module contains all signals.
 from django.dispatch import Signal
 
 
-# Signal that fires when a user is graded (in lms/courseware/grades.py)
+# Signal that fires when a user is graded (in lms/grades/course_grades.py)
 GRADES_UPDATED = Signal(providing_args=["username", "grade_summary", "course_key", "deadline"])
 
 # Signal that fires when a user is awarded a certificate in a course (in the certificates django app)
