+++ setup.py
@@ -52,7 +52,7 @@ setup(
             "hidden_content = lms.djangoapps.course_blocks.transformers.hidden_content:HiddenContentTransformer",
             "course_blocks_api = lms.djangoapps.course_api.blocks.transformers.blocks_api:BlocksAPITransformer",
             "proctored_exam = lms.djangoapps.course_api.blocks.transformers.proctored_exam:ProctoredExamTransformer",
-            "grades = lms.djangoapps.courseware.transformers.grades:GradesTransformer",
+            "grades = lms.djangoapps.grades.transformer:GradesTransformer",
         ],
     }
 )
